requirements.txt
Flask==2.0.1
Pillow==10.0.0
torch==2.5.0
torchvision==0.20.0
Werkzeug==2.0.3
pytesseract==0.3.10
fuzzywuzzy==0.18.0
requests==2.28.1
Flask-PyMongo 
